requirements.txt
accelerate~=1.8.0
fastapi~=0.104.0
google-generativeai~=0.8.0
msgpack~=1.0.0
openai~=1.93.0
pillow~=11.2.0
timm~=1.0.0
transformers~=4.53.0
uvicorn~=0.24.0
websockets~=12.0
